requirements.txt
playwright
python-dotenv
Flask